Moves environments.rb to config/environment.rb
